Add status amount to player view
Displays the amount/value for active statuses in brackets on the player view
for statuses that have a value.

Changes to status amounts were previously not added to the store and only
handled in the Status component. This commit changes performCreatureUpdate()
to always add a status or change its amount when change.status is used; to
remove a status it's now necessary to set change.removeStatus.